+++ cognite/neat/rules/importers/_rdf/_inference2rules.py
@@ -11,7 +11,7 @@ from cognite.neat.issues import IssueList
 from cognite.neat.issues.errors import FileReadError
 from cognite.neat.rules._shared import ReadRules
 from cognite.neat.rules.importers._base import BaseImporter
-from cognite.neat.rules.models._base import MatchType
+from cognite.neat.rules.models._base_rules import MatchType
 from cognite.neat.rules.models.information import (
     InformationInputRules,
     InformationMetadata,
