--- cognite/neat/rules/models/_types/__init__.py
@@ -1,19 +1,0 @@
-from ._field import (
-    ExternalIdType,
-    NamespaceType,
-    PrefixType,
-    PropertyType,
-    StrListType,
-    StrOrListType,
-    VersionType,
-)
-
-__all__ = [
-    "StrOrListType",
-    "StrListType",
-    "NamespaceType",
-    "PrefixType",
-    "ExternalIdType",
-    "VersionType",
-    "PropertyType",
-]
