+++ cognite/neat/rules/models/asset/_validation.py
@@ -3,7 +3,7 @@ from typing import cast
 
 from cognite.neat.issues import IssueList
 from cognite.neat.issues.errors import NeatValueError, PropertyDefinitionError
-from cognite.neat.rules.models._base import SheetList
+from cognite.neat.rules.models._base_rules import SheetList
 from cognite.neat.rules.models.asset._rules import AssetProperty, AssetRules
 from cognite.neat.rules.models.entities import AssetEntity, AssetFields, ClassEntity
 from cognite.neat.rules.models.information._validation import InformationPostValidation
