+++ cognite/neat/rules/models/information/_validation.py
@@ -4,7 +4,7 @@ from typing import cast
 
 from cognite.neat.issues import IssueList
 from cognite.neat.issues.errors import NeatValueError, ResourceNotDefinedError
-from cognite.neat.rules.models._base import DataModelType, SchemaCompleteness
+from cognite.neat.rules.models._base_rules import DataModelType, SchemaCompleteness
 from cognite.neat.rules.models.entities import ClassEntity, EntityTypes, UnknownEntity
 from cognite.neat.utils.rdf_ import get_inheritance_path
